Refactor for opt-in defmt
